removed most get/post/put/delete methods from BackdoorControl. JRADEV-14344
these methods did not do anything anymore and it is better to use the
WebResource methods directly. we don't want BackdoorControl to grow as
large as FuncTestBase in JIRA.

I've left the get(WebResource) method, which is used as a shortcut to
WebResource.get(String.class) and used all over the shop for mutative
operations (the convenience of being able to hit a backdoor with your
browser outweighs the ugliness of using GET for mutative operations).